# docs/swagger.yaml
definitions:
  dto.UserGetOtpRequest:
    properties:
      phone_number:
        maxLength: 12
        minLength: 9
        type: string
    required:
    - phone_number
    type: object
  dto.UserSignUpInput:
    properties:
      password:
        maxLength: 64
        minLength: 8
        type: string
      password_verification:
        maxLength: 64
        minLength: 8
        type: string
      phone_number:
        maxLength: 13
        type: string
    required:
    - password
    - password_verification
    - phone_number
    type: object
info:
  contact: {}
paths:
  /api/v1/users/get-otp:
    post:
      consumes:
      - application/json
      description: user gets otp which need for registration
      parameters:
      - description: get otp info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UserGetOtpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: ok
          schema:
            type: string
        "404":
          description: ok
          schema:
            type: string
        "500":
          description: ok
          schema:
            type: string
        default:
          description: ok
          schema:
            type: string
      summary: Get Otp Code
      tags:
      - Users
  /api/v1/users/sign-up:
    post:
      consumes:
      - application/json
      description: create user account
      parameters:
      - description: sign up info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UserSignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
        "400":
          description: ok
          schema:
            type: string
        "404":
          description: ok
          schema:
            type: string
        "500":
          description: ok
          schema:
            type: string
        default:
          description: ok
          schema:
            type: string
      summary: User SignUp
      tags:
      - Users
swagger: "2.0"